Create package hierarchy
