+++ curvesim/__init__.py
@@ -1,0 +1,3 @@
+from .sim import autosim
+
+__all__ = ["autosim"]
