+++ curvesim/pool/__init__.py
@@ -1,0 +1,4 @@
+from .data import getpool, pooldata
+from .pool import pool
+
+__all__ = ["pool", "getpool", "pooldata"]
